Reorg for a bit of cleanliness, and working unittests.
