+++ app.py
@@ -1,33 +1,9 @@
 #!/usr/bin/python
 
 # Flask is used to create a somewhat lightweight listening server
-from flask import Flask
-from requests import get
+from proxy.init import spawn_proxy
 
+app = spawn_proxy()
 
-myproxy = Flask('__name__')
-
-
-class Proxy:
-
-    # Quick health check override
-    @myproxy.route('/healthcheck', methods=['GET'])
-    def health(self):
-        return "OK"
-
-    # Let's not spam google if we don't get a query, and return a Bad Request.
-    @myproxy.route('/', methods=['GET'])
-    def empty(self):
-        return "Empty search string", 400
-
-    # This is a very dumb proxy, we're only doing GET.
-    @myproxy.route('/<path:req>', methods=['GET'])
-    def proxy(req):
-        # We're only going to google here, so let's just keep it in the proxy settings for now.
-        target = 'https://www.google.com/'
-        return get(f'{target}/search?q={req}').content
-
-    myproxy.run(host='0.0.0.0', port=8080)
-
-
-Proxy()
+if __name__ == '__main__':
+    app.run(host='0.0.0.0', port=8080)
